Remove _udevrulesdir global macro
Patch 271dd9bca1288e732f8d80d98a3d022e5c466bcc changed the definition
of the "_udevrulesdir" global macro to use %{_libdir}, and with that
added a regression that prevents the udev rule "12-vdsm-lvm" from
being installed, which causes our logical volumes to be own by root.
This leads to a situation where libvirt cannot access them at all, so
for example, a Vm with a block domain cannot be started.

This patch removes this macro as it's already defined in macros.systemd.

Change-Id: I4c55297cdf08c5bc2484792a171880ee2fef25da
Signed-off-by: Idan Shaby <ishaby@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/62196
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Tested-by: Eyal Edri <eedri@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Nir Soffer <nsoffer@redhat.com>