vdsm: drop the warnings module configuration
We used to configure the warnings module, because we used
this module to report the inconsistencies between the actual
request/responses and the schema.
We don't do this anymore, so we don't need anymore to configure
the warnings module in our application code.

We will still use the warnings module to warn the clients
of deprecated APIs.
Application using the Vdsm libraries may will be in charge
of configuring the warnings filter if they want to show warnings.

Change-Id: I791882a2dd8ba2dda9135b087bca33610db8a20d
Related-To: https://bugzilla.redhat.com/1364149
Bug-Url: https://bugzilla.redhat.com/1369822
Backport-To: 4.0
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/62212
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Irit Goihman <igoihman@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/64517