virt: vmchannels: use shorter name for thread
We started to use concurrent.thread() in
vmchannels.Listener, in order to have the system
name for the thread, to make it easier to track using
system utilities (e.g. htop).

We need to use a short name, because it needs to be long
15 ASCII characters maxiumum.
We need to make it lowercase, for consistency with other names.

This patch implements those changes.

Change-Id: I19009c85f007ba21e205c94e935715eb28ba004a
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/62276
Continuous-Integration: Jenkins CI
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>