net: Adding the 'link' package with an iface module
The link package is to be used for all iface, bond and other related actions.
Its main purpose is to provide an api to access these devices and their
attributes.

The plan is to add drivers for the underlying devices access.

This patch adds the first link module: iface, which present an api to
iface related actions (link up, link down, etc)

Change-Id: Icb723b8d893575ef14e71cacb6b6e391a6a84831
Signed-off-by: Edward Haas <edwardh@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1379115
Reviewed-on: https://gerrit.ovirt.org/62827
Continuous-Integration: Jenkins CI
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/64389