Adding build-artifacts-manual.sh script
Adding build-artifacts-manual.sh script and related
packages and repos files.
build-artifacts-manual.sh is used for creating rpms from an
existing tarball. It assumes that the tarball already exists
in the projects directory and builds srpms and rpms, and
moves the files to exported_artifacts directory.
For more info: https://gerrit.ovirt.org/#/c/62811/

Change-Id: Ideae96f685caa7ece1295e28af65b0317244fa10
Signed-off-by: Sharon Naftaly <snaftaly@redhat.com>