sdm: Pass the real hostid to copy_data
We have been passing the hostid as None to copy_data but the correct
value will need to be passed in order to support volume leases.  Since
we use the same hostid for all storage domains and the that hostid is
the same as the id attribute of the connected storage pool we can add a
simple helper to supply the proper hostid to sdm verbs.  This will need
to be changed if we remove storage pools or if we opt to use different
hostid values for each domain.

Change-Id: I367576fca126be695d7617477f92110941620a3b
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/62576
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI