Adding the SDM.move_domain_device verb
This patch defines the SDM.move_domain_device verb that will
be used to move data from a block domain device to other devices
that are part of the domain.
This verb will be used currently prior to reducing a device from a
storage domain.

Currently throws UnsupportedOperation exception till BZ 1403836 will be
resolved.

Change-Id: I74183d13061d114a59da23874c86186457046e94
Signed-off-by: Liron Aravot <laravot@redhat.com>