lib: api: allow to return prebuilt responses
We want to switch to a saner approach on which responses are built
on the API layer - no longer on the application layer.
So, application layer can just raise a proper Exception, and the
API layer will catch and create the appropriate error response.

But we have a huge pile of code to translate, so we need a gradual
approach.
This patch extends common.api to passthrough a prebuilt response,
so we can migrate our application code incrementally.

Change-Id: I4a1c6ab93271e2c5619483446403068ddd95ecb4
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61474
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI