lib: response: helper to detect valid responses
We should build responses in the API layer, and the responses
should be built from what the application layer gives back,
be that return values, or exceptions

We need a gradual upgrade path due to huge amount of legacy code,
so we need to handle the response objects built in the
application layer. For the short/mid-term, the API layer
needs to cope with both flows, modern and old.

So we need an utility to detect which kind the response
the application layer returns to the API layer.

This patch adds this utility, which will be erased once
the transition is complete, with the other scaffolding.

Change-Id: Ifab6a25bea1b4a187d8425275e86bdb2fecf4c7d
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63760
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>