build: require only the same version, not release
The various vdsm subpackages where requiring version-release.
The `release` value, however, depends on the build, hence it
can get out of sync when building on different architectures
(e.g. x86_64, ppc64), so we can end up depending on packaged
which does not exists in the repository.

We recently had the same issue on ioprocess
(https://gerrit.ovirt.org/#/c/62464/1), so we are borrowing
the same solution: we set dependencies only on versions,
ignorning the release.

Like ioprocess, Vdsm must be backward compatible within
the same release, so nothing is lost from the safety side.

Change-Id: Ibf5366b0d38471114ddd3ad82e1402ebd308338e
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/62672
Continuous-Integration: Jenkins CI
Reviewed-by: Irit Goihman <igoihman@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>