tests: mom: create the socket under tmpdir
Some tests may fail under automation if Vdsm
is built and its test are run under a too nested
directory.
This because the absolute path of the UNIX domain
socket becomes too long.

To fix this, we create the socket under a temporary
directory.
This also makes cleanup a little easier.

Change-Id: Ib41253a58e8abadeb99d57072ace75461a8e078b
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/62779
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Roman Mohr <rmohr@redhat.com>
Reviewed-by: Martin Sivák <msivak@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>