tests: split vmfakecon out of vmfakelib
vmfakelib contains both a fake ClientIF and a fake libvirtconnection.
Unlike ClientIF, libvirtconnection is a smaller entity that is easier to
move and possible to import from lib/vdsm as well as from py3.

This patch moves libvirtconnection into another module so it is
usable by py3 tests.

Change-Id: Id393cf07c0de8ea06dceec3d79e5d618f79fa6cc
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63197
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI