net: Link setup module - includes bond setup logic.
Introducing the bond setup module.
It includes the setup logic for Bond devices, translating the setup
request to the actions needed to apply them.

It uses the bond interface for applying the setup.

The bond setup is using the bond transaction context, allowing
cleanup to occur in case of a failure.
Leaving the bond in a controlled state
(mainly, trying to avoid defining only a partial bond).

Change-Id: Ib8f01a401cb1b96e357bc462e528a2a547c59c24
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/62831
Continuous-Integration: Jenkins CI
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>