virt: Initial support for post-copy migration
This patch introduces the basics of switching to post-copy migration
when present in the convergence schedule.  It is an initial step towards
post-copy migration feature in oVirt.  Further changes, most notably
support in Engine, are needed.

As for Vdsm itself, there are several problems to solve in the followup
patches, especially:

- Destination actions that are be performed after pre-copy migration
  finishes should be also performed when pre-copy migration switches to
  post-copy.  The VM is already running on the destination at the moment
  so it must be properly set up immediately, without waiting for
  migration completion.

- The source host doesn't provide (real) stats after switching to
  post-copy mode.  The stats are available on the destination host, with
  the exception of migration progress and migration downtime, after
  switching to post-copy so they must be retrieved from there.
  Additionally, downtime is reported in a bit different way than for
  pre-copy migrations.  We must inform Engine what's happening so that
  it knows where and how to get all the data.

There are some circumstances under which post-copy migration doesn't
work, such as when using (non-transparent) huge pages or RDMA.  If
libvirt can detect it, it may fail already at the start of the migration
when VIR_MIGRATE_POSTCOPY flag is present.

We don't want to block migrations completely if something like that
happens unexpectedly.  So we add VIR_MIGRATE_POSTCOPY flag only when
post-copy migration is actually requested.  This way, if the migration
can't be started with post-copy, another schedule may be attempted.

Please note that (correctly working) post-copy migration requires recent
QEMU and sufficiently new libvirt.  The crucial dependencies are handled
in another patch.

Change-Id: I4c9911f47331120a1b78326044a4949abf35d5fe
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1354343