spec: Require ioprocess 0.15.2
The first time IOProcess.truncate() or IOProcess.touch() is called,
ioprocess keep the file open and will never close it.

Typically when using with vdsm, ioprocess will keep the
__DIRECT_IO_TEST__ file open on shared storage, since the first thing
vdsm does is touching this file.

Keeping open files on shared storage may prevent activation of a storage
domain after it was deactivated.

This is a 3.6 only patch, 4.0 already requires ioprocess 0.16.1
containing this fix.

Change-Id: Ia1fba8f5bfb84e9ff60478e36a09f1690ca6989a
Bug-Url: https://bugzilla.redhat.com/1370564
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63038
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>