virt: Fetch only minimum VM stats during post-copy migration
When a migration switches to post-copy mode, the VM starts running on
the destination.  Libvirt no longer provides actual stats on the source,
with the exception of progress, which is still available on the source
and only on the source.

This patch stops reporting stats other than progress from libvirt during
post-copy migration.

Change-Id: Iac3da08b0816e7a9a247969147c17b15518c7624
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1354343