misc: Remove log skipping infrastructure
We had complex and expensive code for logging the caller of
resource manager requests. It was originally introduced in order to 
help debug culprits of deadlocks. Such deadlocks have not been seen for
years.

Other complex code was added so that logging in a common helper function
would rather use the line number of the helper function's caller, since
the latter is the more useful for debugging (rhbz#572050). @logskip was 
ever used very seldom (though using it more might have improved our 
code).

The code is not very important, yet it too complex and a pain to
maintain or port to Python 3. Let us drop it now.

This feature is not needed and I don't want
to maintain this code on both Python 2 and 3.

Change-Id: I714104458fb3794237c46548068b0842301404e5
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63267
Continuous-Integration: Jenkins CI
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>