virt: Log post-copy migration cancellation failure
A VM in post-copy migration is split between the source and the
destination, with inconsistent memory state on both the ends of the
migration.  Stopping the migration in that phase is not a good idea, it
loses the VM and leaves it in the pause state on the source and may lead
to data loss.

For those reasons libvirt prevents aborting a migration in the post-copy
mode.  Engine prevents canceling post-copy migrations too, but it may
still request aborting them during the moments it's not (yet) aware
about the post-copy mode.  In such a situation, the cancel request fails
and we should at least log the reason.

Change-Id: I670ab48ed3fe69582646675d8b683bcdb3cfa038
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1354343