net: Bond - adding a transaction context.
The bond transaction context is aimed to keep bonds in a controlled
state when adding, editing and removing bonds.

Specifically, it is targeted to handle the following scenario:
- Create bond99 - Success
- Add slave88   - Success
- Add slave-bad - Failure
In the creation of bond99, adding a slave failed, leaving a partial bond
configured.
When setting the bond in a transaction, bond99 will be removed before
re-raising the exception.

Change-Id: If8d015ed235c3620c04131e57ad079a2ec1a0ecc
Bug-Url: https://bugzilla.redhat.com/1379115
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63443
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/64402