virt: Start all VM related stuff after switching to post-copy migration
On the migration destination, we perform running-VM actions only after
the migration completes.  But when the migration switches to post-copy
mode, the VM starts fully running on the destination.  So switching to
post-copy should be considered as the actual start of the VM and
appropriate actions should be performed.

This patch implements unblocking the start up actions after switching to
post-copy mode.

The only exception is VM status, which must still signal migration until
the migration completes.

Change-Id: Ic8b17e58f63cbd9db09e4420871a562eaa0b3f3d
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1354343