Live Merge: Refresh base volume before live merge
When live merging raw base volume, engine extends the base volume
if it is smaller than the top volume. However, on the host running
the vm, the lv is already active and does not reflect the new size
until we refresh it.  During the merge we don't know whether base
volume extended, so we always refresh it.

Change-Id: I787d6854e780035b09e4f09d71ca776342dff5be
Bug-Url: https://bugzilla.redhat.com/1367281
Signed-off-by: Ala Hino <ahino@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63454
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Nir Soffer <nsoffer@redhat.com>