jsonrpcvdscli: fix Host.getVMList command result
Host.getVMList and Host.getFullVMList are similar commands that differ in
their verbosity level. Their result should be represented as a value inside
a dictionary with 'vmList' key. Instead of that, it was represented as
a list and during the creation of JsonRPCResponse it was re constructed
inside a dictionary holding the result inside 'items' key.

Change-Id: I28f39417b0006957670bb6ccbdb673f2e0f89043
Signed-off-by: Irit Goihman <igoihman@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63408
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>