net test: test_ip_info - Wait for IP setting to take affect
test_ip_info is failing randomly on CI.
The suspected problem is the time it takes from the action of adding an
IP to the actual kernel apply.

In this patch, on each IP add, the test waits for an event of 'done'.

Failing test showed that the 198.51.100.11/32 is missing.

Change-Id: Ia728db9739429df760747b259c488684cc2f378e
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63516
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>