build tests: Require NOSE 1.3.7 and up for running tests
On RHEL7/Centos7 the provided NOSE version is 1.3.0.
CI runs the tests with 1.3.7.

To be consistent and avoid different behaviours, assure that the tests
are running with a minimum nose version of 1.3.7.

Specifically, between 1.3.0 and 1.3.7 a bug has been resolved regarding
test labeling and its support with test class inheritance.

Change-Id: If79d8624cee1c14a21840e4a08000fc33abb58e5
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63638
Continuous-Integration: Jenkins CI
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Reviewed-by: Irit Goihman <igoihman@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>