virt: Destroy VM after post-copy migration failure
As explained in the source code comment, we don't have currently a
better option than to destroy the VM remnants after a failed post-copy
migration.  This may change in future, if a failed post-copy migration
recovery is available in libvirt/QEMU.

How do we know that a post-copy migration failed?  On the source, the migration
simply fails and the VM either remains in a paused status with a post-copy
reason or it disappears.  On the destination, the VM can end up in the same
states, but there are two more things to consider: 1. we may not be aware that
we are in post-copy (if the corresponding event is not sent from libvirt before
the migration fails); 2. if the VM gets paused, we receive
VIR_DOMAIN_EVENT_SUSPENDED_POSTCOPY_FAILED event.

Disappeared VMs are already handled as general QEMU crashes.  We must
specifically handle the paused case on both the ends, since this is a new
situation and we want to destroy the VM immediately.  We also delete the VM
object on the source, since Engine handles only the destination in post-copy
migrations.

Change-Id: I1918e9afce189c8b3f617766e55afa13f1e153f1
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1354343