lib: api.py: preserve signature in decorator
The api.method decorator needs to preserve the function
or method signature, to let the schema validation keep working.
Moreover, preserving the signature of decorated callables
is a good idea anyway.

In python 3.4+ this is done by default, but on python 2.7
we need to take steps to implement this.
We leverage the existing 'decorator' module which solves
this problem. This module is available in EL/Centos7.

For further reference:
http://pythonhosted.org/decorator/documentation.html

Change-Id: I06205cf180229ea19ffc38a7b88346afc18f13e4
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63759
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>