sos: collect 'nodectl info' output
CEE needs a quick and reliable way to determine release information
from a customer's sosreport. nodectl is installed RHV-H by default
and the output of 'nodectl info' provides a nice summary of the
available images and their boot options as well as the current image/layer
in use. If the command doesn't exist it will be ignored by sos report.

Change-Id: I8360e3812d9a84a592099ba93c583aa6c68a3ed0
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1372093
Signed-off-by: Irit Goihman <igoihman@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63775
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>