net tests: Remove redundant label from func test template.
The new functional test format are using a template test class with two
sublasses, one for each switch type (legacy, ovs).

With NOSE 1.3.0 there was a need to overcome a bug, that did not pass
correctly the NOSE test labels, by labeling the template class in
addition to the subclasses.

As we are now requiring NOSE 1.3.7 for the test run, the template class
tests have been cleaned.

Change-Id: I2f75bb1d6aabceceff65d673ec776cd5b5c939c3
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63931
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>