jobs: Make jobs.stop no-op
When vdsm stops we were calling jobs.stop which cleared the list of
jobs.  Clearing the list is not required for vdsmd because the jobs list
is not persisted across vdsm runs.  We can use the old _clear method for
test cases which do require a way to clear the list.  We'll make stop()
a no-op which has an added benefit of avoiding race conditions between
stop() and job autodeletion.

Change-Id: Ifa1dc49472f429543a5335747c8e7a4b13bb3112
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63937
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>