lib: move exception.py under common
The commit fc732ae broke the rule that code under lib/vdsm/common
should depend only on code under the same subtree.

The good news is that the code moved in fc732ae depends only on
a few modules; one of them is exception.py, which makes sense
to move under common/ anyway.

This patch does that.

Change-Id: I7742aa7dafe0e0e29044536ad19fbddbcab0f420
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63952
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>