lib: move the response.py under common
The commit fc732ae broke the rule that code under lib/vdsm/common
should depend only on code under the same subtree.

The last module we need to move to fix the breakage is 'response'.
Once again, this makes sense to be moved in the common layer anyway.

Change-Id: I6263fb8991a226516ba5d14e909a4a70ea0078d2
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63954
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>