ResourceManagerLock: Disable autoRelease
When acquiring ResourceManagerLocks we do not store a reference to the
resource (prefering instead to use the ResourceManager instance
methods).  Unfortunately the underlying resource has autoRelease
behavior on by default when the object is destroyed.  This is causing
the lock to be released immediately after it is acquired!  Disable
autoRelease to get the correct behavior.

Change-Id: Ie70dca0079f773362bfebe6b9a08677c94314e68
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63967
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>