virt: Distinguish between switching to and entering post-copy migration
Switching to migration post-copy mode has two phases: Asking for
switching to post-copy and actually switching to post-copy.  The
transition period should be short (if it is not then it is a
QEMU/libvirt bug), so we needn't care about it much.

But on various checks we should distinguish between the post-copy
request phase and the post-copy switch phase.  For instance, we should
block canceling migrations very early to avoid timing issues and
possibly canceling a post-copy migration.  On the other hand we should
report a true VM state to Engine, so we can report post-copy only after
the VM actually enters it.

Change-Id: I88dddeab84a609ebd3d5d9139724adff9cdb1352
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1354343