automation: add a repo with a fresh nose
We need nose >= 1.3.7 to test vdsm, but don't have it on el7. One option
discussed is not to run test when building for el7. As a quick-and dirty
fix for upstream, this patch pulls nose from CentOS virt SIG. It does
NOT solve the problem when building in brew or koji.

Change-Id: Ic862433638be19ea770868b1cd5f0461b2ff6599
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63986
Continuous-Integration: Jenkins CI
Tested-by: Sandro Bonazzola <sbonazzo@redhat.com>
Reviewed-by: Irit Goihman <igoihman@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>