net: Consume ifcfg files that have a non vdsm standard name
VDSM assumes that the ifcfg files which represent network devices are
named in the following format: ifcfg-<dev name>

If the host is set initially by Network Manager, the names of the files
do no correspond to the expected format.

This patch adjusts the ifcfg file name to the mentioned format and
erases any other ifcfg files that correspond to the same device.

This is not a NM friendly patch, it assumes that NM is not active while
setupNetworks is issued.

Change-Id: I0bf70ba936d5de1f17a90742644719216018f674
Bug-Url: https://bugzilla.redhat.com/1367378
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63990
Continuous-Integration: Jenkins CI
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>