virt: Report downtime after a post-copy migration
There are two migration downtime length values provided by libvirt after
a completed migration: `downtime' and `downtime_net'.  Ideally,
`downtime' provides the complete downtime length, but only under the
assumption that the clocks on the hosts are perfectly synchronized.  If
they are not, the value may be screwed by the clock differences.

For this reason, we report `downtime_net' for pre-copy migrations, which
includes only the time needed to transfer the data (and not other
overheads), but is correct even when the clocks are not perfectly
synchronized.

`downtime_net' is not reported for post-copy migrations as it is always
zero.  So only `downtime' can be reported for post-copy migrations,
regardless of the potential problem with the clocks.

Engine and other tools reporting the downtime must take into account the
differences of the downtime value in pre-copy and post-copy.

Change-Id: I5f34e1ff577db325cc976431b3570f91833ff815
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1354343