virt: Stop guest agent before switching to post-copy
Post-copy migration pauses the VM on the source.  We don't want to get
spurious log messages from the guest agent about not being able to talk
to the VM or being disconnected from the communication socket.  So we
stop the guest agent before switching to post-copy mode.

If switching to post-copy fails (which should rarely happen), we start
the agent again.  Starting the agent may fail, e.g. in case the VM got
destroyed in the meantime.  So we catch the contingent exception.

Change-Id: Ibd855ee36847b293009ac35d8d03b01521aaace8
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1354343