net: Split OVS setup transaction to adjust for bond setup
When using Linux bonds (instead of OVS bonds), the OVS setup transaction
needs to be split into two.
First one for removing networks and a second for adding networks.
In between, the bonds setup is placed.

These three steps contain, in addition to the actual devices setup, the
acquire and running-config update, allowing proper rollback in case of
setup failure.

This patch, fixes a scenario that fails with the previous patch, where
the Linux bonds have been introduced in the setup flow.
The scenario: Move a network nic from a network to a bond slave.

Change-Id: I0bd351e755be9d5dcaf5a05e5b79345e6f7bcfe8
Bug-Url: https://bugzilla.redhat.com/1379115
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/64118
Continuous-Integration: Jenkins CI
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/64412