virt: Prevent monitoring of VMs in post-copy
Migrating VMs in post-copy mode on the source end are paused.  It makes
no sense to monitor them and we don't want to get error messages in the
log or experience other problems.  So we skip VMs in that state in
periodic operations.

We also skip the VMs that just successfully initiated switching to
post-copy mode.  Under normal circumstances, they should enter post-copy
mode shortly afterwards, so it's better to skip them immediately.

We still can't guarantee (without excessive locking) that VM doesn't
switch to post-copy (and gets paused), so we handle the corresponding
contingent failure as well.

Note that we still gather stats, since we need to know the migration
progress.

Change-Id: I010d3311ce497081720a871e2a3c715f53dcb299
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1354343