virt: Destroy VM after a successful post-copy migration
When we switch to post-copy migration, Engine starts monitoring the
destination (once it receives the appropriate notification).  After the
migration succeeds, Engine doesn't handle the source anymore and doesn't need
information about VM status from it.  So when the post-copy migration finishes,
we no longer keep the VM object and destroy it immediately.

Change-Id: Iff88aa0aaa88b34167bb671a56f2cc09c122ab6f
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1354343