storage: Move detect_format to new workarounds module
The function Image._detect_format is a workaround for a bug where the
vdsm metadata disagrees with qemuimg about the format of a volume.
While the underlying bug has been fixed, we still need to maintain this
code in order to prevent problems with preexisting affected volumes.

For reusability and maintainability, move this function to a new module
called workarounds.py and add some basic unit tests.  This code will be
reused by sdm.copy_data in a followup patch.

Change-Id: If4da9d2c16679f99b55438d7336d0cfb27429c12
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/64229
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Tested-by: Nir Soffer <nsoffer@redhat.com>