logging: Unify SimpleLogAdapter format
SimpleLogAdapter was prefixing the log message with context info, using
"::" as message separator. This made sense when the log was using "::"
as field separator, but now we are using spaces.

Additionally the old format use strange quoting and no spacing between
pairs, making the format harder to read:

    task=`xxxyyy`res=`foo.bar.baz`::Message text ...

The new format fixes these issues and enclose the context info in
parenthesis to match the new log format:

    (task='xxxyyy', res='foo.bar.baz') Message text ...

This matches the format we used lately in many logs for adding message
details.

Here is an example line using the format:

2016-09-23 02:21:57,965 INFO  (Thread-11) [virt.vm] (vmId='45b59cae-f2a0-4d83-8da8-b7c0a21fbe14') VM wrapper has started (vm:1751)

Change-Id: Ib4edb64d214e515c9cfc45175f6cfee4340d28de
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/64227
Continuous-Integration: Jenkins CI
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-by: Edward Haas <edwardh@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>