virt: Send migration status events
Migration progress is currently available in stats.  Engine fetches the stats
from the host where a migrating VM is running, which is the source host during
pre-copy migration and the destination host during post-copy migration.
However the progress information is always available only on the source host,
so it's not available to Engine during the post-copy phase.

For this and other reasons Engine would like to receive events about migration
progress and the downtime of a completed migration. We introduce a new event
for that purpose containing that information and additionally information about
the pre-copy or post-copy mode of the migration (since downtime value
interpretation is inherently different in the two cases).

Change-Id: I6e392150e0470b99cccdf07f50345bb9e4d02c44
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>