net: Ignore missing graphics from VM conf on target host
The new VM migration libvirt hook handling, recreates the interface and
graphics devices of the domxml to fit the target host. It does so using
the VM conf data.

Unfortunate, there was a bug in Engine that did not pushed the graphics
configuration to VDSM, leaving VM conf without the display information.
VDSM on the other hand, used a default, listen to all, when this data
has not been received from Engine.

On migration, the domxml includes the graphics section while the target
VM conf does not, causing an exception that fails the action.

This patch silently ignores such scenarios, leaving the source domxml as
is in the graphics section.

Change-Id: I8cc730c6448b1f70500c86b3ab39af21fa23bd5a
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/64300
Continuous-Integration: Jenkins CI
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>